best-practices/ai-services/document-information-extraction/python/requirements.txt
asttokens==3.0.0
certifi==2025.4.26
charset-normalizer==3.4.2
contourpy==1.3.0
cycler==0.12.1
decorator==5.2.1
defusedxml==0.7.1
exceptiongroup==1.3.0
executing==2.2.0
fonttools==4.58.1
fpdf2==2.8.3
idna==3.10
importlib_resources==6.5.2
iniconfig==2.1.0
ipython==8.18.1
jedi==0.19.2
joblib==1.5.1
kiwisolver==1.4.7
matplotlib==3.9.4
matplotlib-inline==0.1.7
numpy==2.0.2
oauthlib==3.2.2
packaging==25.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
pluggy==1.6.0
prompt_toolkit==3.0.51
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.3
PyPDF2==3.0.1
pytest==8.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
requests-oauthlib==1.3.1
sap-business-document-processing==0.4.1
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
stack-data==0.6.3
threadpoolctl==3.6.0
tomli==2.2.1
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
wcwidth==0.2.13
zipp==3.22.0
